requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
boto3==1.15.7
botocore==1.18.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-filter==2.4.0
django-storages==1.10.1
djangorestframework==3.12.1
idna==2.10
importlib-metadata==2.0.0
jmespath==0.10.0
Markdown==3.2.2
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
paramiko==2.7.2
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
storage==0.0.4.3
urllib3==1.25.10
zipp==3.2.0
